+++ CHANGELOG.md
@@ -3,7 +3,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## [] - xxx-xx-xx
+
 ## [3.4.6-rciam] - 2022-05-17
+
+### Added
+
+- Support for CO Department types
 
 ### Fixed